+++ packages/EasyCore/src/Search/ElasticsearchSearchService.php
@@ -47,10 +47,10 @@ final class ElasticsearchSearchService implements SearchServiceInterface
                 'must' => $query,
                 'filter' => [
                     [
-                        'terms' => [$accessTokensProperty => $accessTokens ?? ['anonymous']]
-                    ]
-                ]
-            ]
+                        'terms' => [$accessTokensProperty => $accessTokens ?? ['anonymous']],
+                    ],
+                ],
+            ],
         ];
 
         return $this->client->search([