+++ packages/EasyLogging/src/Bridge/Laravel/config/easy-logging.php
@@ -1,4 +1,5 @@
 <?php
+
 declare(strict_types=1);
 
 use EonX\EasyLogging\Interfaces\LoggerFactoryInterface;
@@ -8,4 +9,9 @@ return [
      * Define the default channel name for the application.
      */
     'default_channel' => LoggerFactoryInterface::DEFAULT_CHANNEL,
+
+    /**
+     * Enable/Disable the default stream handler.
+     */
+    'stream_handler' => true,
 ];