+++ packages/EasyLogging/src/Bridge/Symfony/Resources/config/services.php
@@ -13,14 +13,14 @@ use function Symfony\Component\DependencyInjection\Loader\Configurator\tagged_it
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
-    $services
-        ->defaults()
+    $services->defaults()
         ->autowire()
         ->autoconfigure();
 
     $services
         ->set(LoggerFactoryInterface::class, LoggerFactory::class)
-        ->args(['%' . BridgeConstantsInterface::PARAM_DEFAULT_CHANNEL . '%'])
+        ->arg('$defaultChannel', '%' . BridgeConstantsInterface::PARAM_DEFAULT_CHANNEL . '%')
+        ->arg('$loggerClass', '%' . BridgeConstantsInterface::PARAM_LOGGER_CLASS . '%')
         ->call('setHandlerConfigProviders', [tagged_iterator(BridgeConstantsInterface::TAG_HANDLER_CONFIG_PROVIDER)])
         ->call('setLoggerConfigurators', [tagged_iterator(BridgeConstantsInterface::TAG_LOGGER_CONFIGURATOR)])
         ->call(
@@ -29,9 +29,10 @@ return static function (ContainerConfigurator $containerConfigurator): void {
         );
 
     $services
-        ->set(LoggerInterface::class)
+        ->set('easy_logging.logger', '%' . BridgeConstantsInterface::PARAM_LOGGER_CLASS . '%')
         ->factory([ref(LoggerFactoryInterface::class), 'create'])
         ->args(['%' . BridgeConstantsInterface::PARAM_DEFAULT_CHANNEL . '%']);
 
-    $services->alias('logger', LoggerInterface::class);
+    $services->alias('logger', 'easy_logging.logger');
+    $services->alias(LoggerInterface::class, 'logger');
 };