+++ packages/EasyLogging/tests/Stubs/SelfProcessorConfigProviderStub.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EonX\EasyLogging\Tests\Stubs;
+
+use EonX\EasyLogging\Config\AbstractSelfProcessorConfigProvider;
+
+final class SelfProcessorConfigProviderStub extends AbstractSelfProcessorConfigProvider
+{
+    /**
+     * @param mixed[] $records
+     *
+     * @return mixed[]
+     */
+    public function __invoke(array $records): array
+    {
+        return $records;
+    }
+}