+++ packages/EasyStandard/tests/Rector/ExplicitBoolCompareRector/ExplicitBoolCompareRectorTest.php
@@ -7,6 +7,7 @@ namespace EonX\EasyStandard\Tests\Rector\ExplicitBoolCompareRector;
 use EonX\EasyStandard\Rector\ExplicitBoolCompareRector;
 use Iterator;
 use Rector\Core\Testing\PHPUnit\AbstractRectorTestCase;
+use Symplify\SmartFileSystem\SmartFileInfo;
 
 /**
  * @covers \EonX\EasyStandard\Rector\StrictInArrayRector
@@ -16,9 +17,7 @@ use Rector\Core\Testing\PHPUnit\AbstractRectorTestCase;
 final class ExplicitBoolCompareRectorTest extends AbstractRectorTestCase
 {
     /**
-     * Provides test examples.
-     *
-     * @return Iterator<array>
+     * @return Iterator<\Symplify\SmartFileSystem\SmartFileInfo>
      */
     public function provideData(): Iterator
     {
@@ -26,22 +25,15 @@ final class ExplicitBoolCompareRectorTest extends AbstractRectorTestCase
     }
 
     /**
-     * Tests Rector rule.
-     *
      * @dataProvider provideData()
      */
-    public function testRule(string $file): void
+    public function testRule(SmartFileInfo $fileInfo): void
     {
-        $this->doTestFile($file);
+        $this->doTestFileInfo($fileInfo);
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    protected function getRectorsWithConfiguration(): array
+    protected function getRectorClass(): string
     {
-        return [
-            ExplicitBoolCompareRector::class => [],
-        ];
+        return ExplicitBoolCompareRector::class;
     }
 }