+++ packages/EasyStandard/tests/Sniffs/Classes/RequireStrictDeclarationSniff/RequireStrictDeclarationSniffTest.php
@@ -1,0 +1,34 @@
+<?php
+
+declare(strict_types=1);
+
+namespace EonX\EasyStandard\Tests\Sniffs\Classes\RequireStrictDeclarationSniff;
+
+use EonX\EasyStandard\Sniffs\Classes\RequireStrictDeclarationSniff;
+use Iterator;
+use Symplify\EasyCodingStandardTester\Testing\AbstractCheckerTestCase;
+use Symplify\SmartFileSystem\SmartFileInfo;
+
+final class RequireStrictDeclarationSniffTest extends AbstractCheckerTestCase
+{
+    /**
+     * @return Iterator<int, SmartFileInfo|int>
+     */
+    public function providerTestSniff(): iterable
+    {
+        yield [new SmartFileInfo(__DIR__ . '/Fixture/RequireStrictDeclarationSniffTest.php.inc'), 1];
+    }
+
+    /**
+     * @dataProvider providerTestSniff()
+     */
+    public function testSniff(SmartFileInfo $smartFileInfo, int $expectedErrorCount): void
+    {
+        $this->doTestFileInfoWithErrorCountOf($smartFileInfo, $expectedErrorCount);
+    }
+
+    protected function getCheckerClass(): string
+    {
+        return RequireStrictDeclarationSniff::class;
+    }
+}